Move serialize to scylla-cql-core
- No tests
- zigzag / varint copied, because tied to tests and deser. Will remove
later.
- Tests later, and all items from modules re-exported, to prove to bad
items moved. Used LSP to expand list from *.
