+++ scylla-cql-core/src/frame/types.rs
@@ -1,6 +1,7 @@
 //! CQL binary protocol in-wire types.
 
 use byteorder::{BigEndian, ReadBytesExt};
+use bytes::BufMut;
 
 use crate::frame::frame_errors::LowLevelDeserializationError;
 
@@ -195,4 +196,329 @@ pub fn read_value<'a>(buf: &mut &'a [u8]) -> Result<RawValue<'a>, LowLevelDeseri
         }
         len => Err(LowLevelDeserializationError::InvalidValueLength(len)),
     }
+}
+
+fn zig_zag_encode(v: i64) -> u64 {
+    ((v >> 63) ^ (v << 1)) as u64
+}
+
+fn zig_zag_decode(v: u64) -> i64 {
+    ((v >> 1) as i64) ^ -((v & 1) as i64)
+}
+
+pub(crate) fn unsigned_vint_encode(v: u64, buf: &mut Vec<u8>) {
+    let mut v = v;
+    let mut number_of_bytes = (639 - 9 * v.leading_zeros()) >> 6;
+    if number_of_bytes <= 1 {
+        return buf.put_u8(v as u8);
+    }
+
+    if number_of_bytes != 9 {
+        let extra_bytes = number_of_bytes - 1;
+        let length_bits = !(0xff >> extra_bytes);
+        v |= (length_bits as u64) << (8 * extra_bytes);
+    } else {
+        buf.put_u8(0xff);
+        number_of_bytes -= 1;
+    }
+    buf.put_uint(v, number_of_bytes as usize)
+}
+
+pub(crate) fn unsigned_vint_decode(buf: &mut &[u8]) -> Result<u64, std::io::Error> {
+    let first_byte = buf.read_u8()?;
+    let extra_bytes = first_byte.leading_ones() as usize;
+
+    let mut v = if extra_bytes != 8 {
+        let first_byte_bits = first_byte & (0xffu8 >> extra_bytes);
+        (first_byte_bits as u64) << (8 * extra_bytes)
+    } else {
+        0
+    };
+
+    if extra_bytes != 0 {
+        v += buf.read_uint::<BigEndian>(extra_bytes)?;
+    }
+
+    Ok(v)
+}
+
+pub(crate) fn vint_encode(v: i64, buf: &mut Vec<u8>) {
+    unsigned_vint_encode(zig_zag_encode(v), buf)
+}
+
+#[cfg_attr(not(test), expect(unused))]
+pub(crate) fn vint_decode(buf: &mut &[u8]) -> Result<i64, std::io::Error> {
+    unsigned_vint_decode(buf).map(zig_zag_decode)
+}
+
+#[test]
+fn zig_zag_encode_test() {
+    assert_eq!(zig_zag_encode(0), 0);
+    assert_eq!(zig_zag_encode(-1), 1);
+    assert_eq!(zig_zag_encode(1), 2);
+    assert_eq!(zig_zag_encode(-2), 3);
+    assert_eq!(zig_zag_encode(2), 4);
+    assert_eq!(zig_zag_encode(-3), 5);
+    assert_eq!(zig_zag_encode(3), 6);
+}
+
+#[test]
+fn zig_zag_decode_test() {
+    assert_eq!(zig_zag_decode(0), 0);
+    assert_eq!(zig_zag_decode(1), -1);
+    assert_eq!(zig_zag_decode(2), 1);
+    assert_eq!(zig_zag_decode(3), -2);
+    assert_eq!(zig_zag_decode(4), 2);
+    assert_eq!(zig_zag_decode(5), -3);
+    assert_eq!(zig_zag_decode(6), 3);
+}
+
+#[test]
+fn unsigned_vint_encode_and_decode_test() {
+    let unsigned_vint_encoding = vec![
+        (0, vec![0]),
+        (1, vec![1]),
+        (2, vec![2]),
+        ((1 << 2) - 1, vec![3]),
+        (1 << 2, vec![4]),
+        ((1 << 2) + 1, vec![5]),
+        ((1 << 3) - 1, vec![7]),
+        (1 << 3, vec![8]),
+        ((1 << 3) + 1, vec![9]),
+        ((1 << 4) - 1, vec![15]),
+        (1 << 4, vec![16]),
+        ((1 << 4) + 1, vec![17]),
+        ((1 << 5) - 1, vec![31]),
+        (1 << 5, vec![32]),
+        ((1 << 5) + 1, vec![33]),
+        ((1 << 6) - 1, vec![63]),
+        (1 << 6, vec![64]),
+        ((1 << 6) + 1, vec![65]),
+        ((1 << 7) - 1, vec![127]),
+        (1 << 7, vec![128, 128]),
+        ((1 << 7) + 1, vec![128, 129]),
+        ((1 << 8) - 1, vec![128, 255]),
+        (1 << 8, vec![129, 0]),
+        ((1 << 8) + 1, vec![129, 1]),
+        ((1 << 9) - 1, vec![129, 255]),
+        (1 << 9, vec![130, 0]),
+        ((1 << 9) + 1, vec![130, 1]),
+        ((1 << 10) - 1, vec![131, 255]),
+        (1 << 10, vec![132, 0]),
+        ((1 << 10) + 1, vec![132, 1]),
+        ((1 << 11) - 1, vec![135, 255]),
+        (1 << 11, vec![136, 0]),
+        ((1 << 11) + 1, vec![136, 1]),
+        ((1 << 12) - 1, vec![143, 255]),
+        (1 << 12, vec![144, 0]),
+        ((1 << 12) + 1, vec![144, 1]),
+        ((1 << 13) - 1, vec![159, 255]),
+        (1 << 13, vec![160, 0]),
+        ((1 << 13) + 1, vec![160, 1]),
+        ((1 << 14) - 1, vec![191, 255]),
+        (1 << 14, vec![192, 64, 0]),
+        ((1 << 14) + 1, vec![192, 64, 1]),
+        ((1 << 15) - 1, vec![192, 127, 255]),
+        (1 << 15, vec![192, 128, 0]),
+        ((1 << 15) + 1, vec![192, 128, 1]),
+        ((1 << 16) - 1, vec![192, 255, 255]),
+        (1 << 16, vec![193, 0, 0]),
+        ((1 << 16) + 1, vec![193, 0, 1]),
+        ((1 << 17) - 1, vec![193, 255, 255]),
+        (1 << 17, vec![194, 0, 0]),
+        ((1 << 17) + 1, vec![194, 0, 1]),
+        ((1 << 18) - 1, vec![195, 255, 255]),
+        (1 << 18, vec![196, 0, 0]),
+        ((1 << 18) + 1, vec![196, 0, 1]),
+        ((1 << 19) - 1, vec![199, 255, 255]),
+        (1 << 19, vec![200, 0, 0]),
+        ((1 << 19) + 1, vec![200, 0, 1]),
+        ((1 << 20) - 1, vec![207, 255, 255]),
+        (1 << 20, vec![208, 0, 0]),
+        ((1 << 20) + 1, vec![208, 0, 1]),
+        ((1 << 21) - 1, vec![223, 255, 255]),
+        (1 << 21, vec![224, 32, 0, 0]),
+        ((1 << 21) + 1, vec![224, 32, 0, 1]),
+        ((1 << 22) - 1, vec![224, 63, 255, 255]),
+        (1 << 22, vec![224, 64, 0, 0]),
+        ((1 << 22) + 1, vec![224, 64, 0, 1]),
+        ((1 << 23) - 1, vec![224, 127, 255, 255]),
+        (1 << 23, vec![224, 128, 0, 0]),
+        ((1 << 23) + 1, vec![224, 128, 0, 1]),
+        ((1 << 24) - 1, vec![224, 255, 255, 255]),
+        (1 << 24, vec![225, 0, 0, 0]),
+        ((1 << 24) + 1, vec![225, 0, 0, 1]),
+        ((1 << 25) - 1, vec![225, 255, 255, 255]),
+        (1 << 25, vec![226, 0, 0, 0]),
+        ((1 << 25) + 1, vec![226, 0, 0, 1]),
+        ((1 << 26) - 1, vec![227, 255, 255, 255]),
+        (1 << 26, vec![228, 0, 0, 0]),
+        ((1 << 26) + 1, vec![228, 0, 0, 1]),
+        ((1 << 27) - 1, vec![231, 255, 255, 255]),
+        (1 << 27, vec![232, 0, 0, 0]),
+        ((1 << 27) + 1, vec![232, 0, 0, 1]),
+        ((1 << 28) - 1, vec![239, 255, 255, 255]),
+        (1 << 28, vec![240, 16, 0, 0, 0]),
+        ((1 << 28) + 1, vec![240, 16, 0, 0, 1]),
+        ((1 << 29) - 1, vec![240, 31, 255, 255, 255]),
+        (1 << 29, vec![240, 32, 0, 0, 0]),
+        ((1 << 29) + 1, vec![240, 32, 0, 0, 1]),
+        ((1 << 30) - 1, vec![240, 63, 255, 255, 255]),
+        (1 << 30, vec![240, 64, 0, 0, 0]),
+        ((1 << 30) + 1, vec![240, 64, 0, 0, 1]),
+        ((1 << 31) - 1, vec![240, 127, 255, 255, 255]),
+        (1 << 31, vec![240, 128, 0, 0, 0]),
+        ((1 << 31) + 1, vec![240, 128, 0, 0, 1]),
+        ((1 << 32) - 1, vec![240, 255, 255, 255, 255]),
+        (1 << 32, vec![241, 0, 0, 0, 0]),
+        ((1 << 32) + 1, vec![241, 0, 0, 0, 1]),
+        ((1 << 33) - 1, vec![241, 255, 255, 255, 255]),
+        (1 << 33, vec![242, 0, 0, 0, 0]),
+        ((1 << 33) + 1, vec![242, 0, 0, 0, 1]),
+        ((1 << 34) - 1, vec![243, 255, 255, 255, 255]),
+        (1 << 34, vec![244, 0, 0, 0, 0]),
+        ((1 << 34) + 1, vec![244, 0, 0, 0, 1]),
+        ((1 << 35) - 1, vec![247, 255, 255, 255, 255]),
+        (1 << 35, vec![248, 8, 0, 0, 0, 0]),
+        ((1 << 35) + 1, vec![248, 8, 0, 0, 0, 1]),
+        ((1 << 36) - 1, vec![248, 15, 255, 255, 255, 255]),
+        (1 << 36, vec![248, 16, 0, 0, 0, 0]),
+        ((1 << 36) + 1, vec![248, 16, 0, 0, 0, 1]),
+        ((1 << 37) - 1, vec![248, 31, 255, 255, 255, 255]),
+        (1 << 37, vec![248, 32, 0, 0, 0, 0]),
+        ((1 << 37) + 1, vec![248, 32, 0, 0, 0, 1]),
+        ((1 << 38) - 1, vec![248, 63, 255, 255, 255, 255]),
+        (1 << 38, vec![248, 64, 0, 0, 0, 0]),
+        ((1 << 38) + 1, vec![248, 64, 0, 0, 0, 1]),
+        ((1 << 39) - 1, vec![248, 127, 255, 255, 255, 255]),
+        (1 << 39, vec![248, 128, 0, 0, 0, 0]),
+        ((1 << 39) + 1, vec![248, 128, 0, 0, 0, 1]),
+        ((1 << 40) - 1, vec![248, 255, 255, 255, 255, 255]),
+        (1 << 40, vec![249, 0, 0, 0, 0, 0]),
+        ((1 << 40) + 1, vec![249, 0, 0, 0, 0, 1]),
+        ((1 << 41) - 1, vec![249, 255, 255, 255, 255, 255]),
+        (1 << 41, vec![250, 0, 0, 0, 0, 0]),
+        ((1 << 41) + 1, vec![250, 0, 0, 0, 0, 1]),
+        ((1 << 42) - 1, vec![251, 255, 255, 255, 255, 255]),
+        (1 << 42, vec![252, 4, 0, 0, 0, 0, 0]),
+        ((1 << 42) + 1, vec![252, 4, 0, 0, 0, 0, 1]),
+        ((1 << 43) - 1, vec![252, 7, 255, 255, 255, 255, 255]),
+        (1 << 43, vec![252, 8, 0, 0, 0, 0, 0]),
+        ((1 << 43) + 1, vec![252, 8, 0, 0, 0, 0, 1]),
+        ((1 << 44) - 1, vec![252, 15, 255, 255, 255, 255, 255]),
+        (1 << 44, vec![252, 16, 0, 0, 0, 0, 0]),
+        ((1 << 44) + 1, vec![252, 16, 0, 0, 0, 0, 1]),
+        ((1 << 45) - 1, vec![252, 31, 255, 255, 255, 255, 255]),
+        (1 << 45, vec![252, 32, 0, 0, 0, 0, 0]),
+        ((1 << 45) + 1, vec![252, 32, 0, 0, 0, 0, 1]),
+        ((1 << 46) - 1, vec![252, 63, 255, 255, 255, 255, 255]),
+        (1 << 46, vec![252, 64, 0, 0, 0, 0, 0]),
+        ((1 << 46) + 1, vec![252, 64, 0, 0, 0, 0, 1]),
+        ((1 << 47) - 1, vec![252, 127, 255, 255, 255, 255, 255]),
+        (1 << 47, vec![252, 128, 0, 0, 0, 0, 0]),
+        ((1 << 47) + 1, vec![252, 128, 0, 0, 0, 0, 1]),
+        ((1 << 48) - 1, vec![252, 255, 255, 255, 255, 255, 255]),
+        (1 << 48, vec![253, 0, 0, 0, 0, 0, 0]),
+        ((1 << 48) + 1, vec![253, 0, 0, 0, 0, 0, 1]),
+        ((1 << 49) - 1, vec![253, 255, 255, 255, 255, 255, 255]),
+        (1 << 49, vec![254, 2, 0, 0, 0, 0, 0, 0]),
+        ((1 << 49) + 1, vec![254, 2, 0, 0, 0, 0, 0, 1]),
+        ((1 << 50) - 1, vec![254, 3, 255, 255, 255, 255, 255, 255]),
+        (1 << 50, vec![254, 4, 0, 0, 0, 0, 0, 0]),
+        ((1 << 50) + 1, vec![254, 4, 0, 0, 0, 0, 0, 1]),
+        ((1 << 51) - 1, vec![254, 7, 255, 255, 255, 255, 255, 255]),
+        (1 << 51, vec![254, 8, 0, 0, 0, 0, 0, 0]),
+        ((1 << 51) + 1, vec![254, 8, 0, 0, 0, 0, 0, 1]),
+        ((1 << 52) - 1, vec![254, 15, 255, 255, 255, 255, 255, 255]),
+        (1 << 52, vec![254, 16, 0, 0, 0, 0, 0, 0]),
+        ((1 << 52) + 1, vec![254, 16, 0, 0, 0, 0, 0, 1]),
+        ((1 << 53) - 1, vec![254, 31, 255, 255, 255, 255, 255, 255]),
+        (1 << 53, vec![254, 32, 0, 0, 0, 0, 0, 0]),
+        ((1 << 53) + 1, vec![254, 32, 0, 0, 0, 0, 0, 1]),
+        ((1 << 54) - 1, vec![254, 63, 255, 255, 255, 255, 255, 255]),
+        (1 << 54, vec![254, 64, 0, 0, 0, 0, 0, 0]),
+        ((1 << 54) + 1, vec![254, 64, 0, 0, 0, 0, 0, 1]),
+        ((1 << 55) - 1, vec![254, 127, 255, 255, 255, 255, 255, 255]),
+        (1 << 55, vec![254, 128, 0, 0, 0, 0, 0, 0]),
+        ((1 << 55) + 1, vec![254, 128, 0, 0, 0, 0, 0, 1]),
+        ((1 << 56) - 1, vec![254, 255, 255, 255, 255, 255, 255, 255]),
+        (1 << 56, vec![255, 1, 0, 0, 0, 0, 0, 0, 0]),
+        ((1 << 56) + 1, vec![255, 1, 0, 0, 0, 0, 0, 0, 1]),
+        (
+            (1 << 57) - 1,
+            vec![255, 1, 255, 255, 255, 255, 255, 255, 255],
+        ),
+        (1 << 57, vec![255, 2, 0, 0, 0, 0, 0, 0, 0]),
+        ((1 << 57) + 1, vec![255, 2, 0, 0, 0, 0, 0, 0, 1]),
+        (
+            (1 << 58) - 1,
+            vec![255, 3, 255, 255, 255, 255, 255, 255, 255],
+        ),
+        (1 << 58, vec![255, 4, 0, 0, 0, 0, 0, 0, 0]),
+        ((1 << 58) + 1, vec![255, 4, 0, 0, 0, 0, 0, 0, 1]),
+        (
+            (1 << 59) - 1,
+            vec![255, 7, 255, 255, 255, 255, 255, 255, 255],
+        ),
+        (1 << 59, vec![255, 8, 0, 0, 0, 0, 0, 0, 0]),
+        ((1 << 59) + 1, vec![255, 8, 0, 0, 0, 0, 0, 0, 1]),
+        (
+            (1 << 60) - 1,
+            vec![255, 15, 255, 255, 255, 255, 255, 255, 255],
+        ),
+        (1 << 60, vec![255, 16, 0, 0, 0, 0, 0, 0, 0]),
+        ((1 << 60) + 1, vec![255, 16, 0, 0, 0, 0, 0, 0, 1]),
+        (
+            (1 << 61) - 1,
+            vec![255, 31, 255, 255, 255, 255, 255, 255, 255],
+        ),
+        (1 << 61, vec![255, 32, 0, 0, 0, 0, 0, 0, 0]),
+        ((1 << 61) + 1, vec![255, 32, 0, 0, 0, 0, 0, 0, 1]),
+        (
+            (1 << 62) - 1,
+            vec![255, 63, 255, 255, 255, 255, 255, 255, 255],
+        ),
+        (1 << 62, vec![255, 64, 0, 0, 0, 0, 0, 0, 0]),
+        ((1 << 62) + 1, vec![255, 64, 0, 0, 0, 0, 0, 0, 1]),
+        (
+            (1 << 63) - 1,
+            vec![255, 127, 255, 255, 255, 255, 255, 255, 255],
+        ),
+        (1 << 63, vec![255, 128, 0, 0, 0, 0, 0, 0, 0]),
+        ((1 << 63) + 1, vec![255, 128, 0, 0, 0, 0, 0, 0, 1]),
+        (u64::MAX, vec![255, 255, 255, 255, 255, 255, 255, 255, 255]),
+    ];
+
+    let mut buf = Vec::new();
+
+    for (v, result) in unsigned_vint_encoding.into_iter() {
+        unsigned_vint_encode(v, &mut buf);
+        assert_eq!(buf, result);
+        let decoded_v = unsigned_vint_decode(&mut buf.as_slice()).unwrap();
+        assert_eq!(v, decoded_v);
+        buf.clear();
+    }
+}
+
+#[test]
+fn vint_encode_and_decode_test() {
+    let mut buf: Vec<u8> = Vec::with_capacity(128);
+
+    let mut check = |n: i64| {
+        vint_encode(n, &mut buf);
+        assert_eq!(vint_decode(&mut buf.as_slice()).unwrap(), n);
+        buf.clear();
+    };
+
+    for i in 0..63 {
+        check((1 << i) - 1);
+        check(1 - (1 << i));
+        check(1 << i);
+        check(-(1 << i));
+        check((1 << i) + 1);
+        check(-1 - (1 << i));
+    }
+    check(i64::MAX);
+    check(-i64::MAX);
+    check(i64::MIN)
 }
