+++ scylla-cql-core/src/serialize/mod.rs
@@ -1,0 +1,54 @@
+#![warn(missing_docs)]
+
+//! Types and traits related to serialization of values to the CQL format.
+
+use std::{error::Error, sync::Arc};
+
+use thiserror::Error;
+
+pub mod batch;
+pub mod row;
+pub mod value;
+pub mod writers;
+
+pub use writers::{CellValueBuilder, CellWriter, RowWriter};
+
+/// An error indicating that a failure happened during serialization.
+///
+/// The error is type-erased so that the crate users can define their own
+/// serialization impls and their errors. As for the impls defined or generated
+/// by the driver itself, the following errors can be returned:
+///
+/// - [`row::BuiltinSerializationError`] is returned when serialization of
+///   one of types with an impl built into the driver fails. It is also returned
+///   from impls generated by the `SerializeRow` macro.
+/// - [`value::BuiltinSerializationError`] is analogous to the above but is
+///   returned from [`SerializeValue::serialize`](value::SerializeValue::serialize)
+///   instead both in the case of builtin impls and impls generated by the
+///   `SerializeValue` macro. It won't be returned by the `Session` directly,
+///   but it might be nested in the [`row::BuiltinSerializationError`].
+#[derive(Debug, Clone, Error)]
+#[error("SerializationError: {0}")]
+pub struct SerializationError(Arc<dyn Error + Send + Sync>);
+
+impl SerializationError {
+    /// Constructs a new `SerializationError`.
+    #[inline]
+    pub fn new(err: impl Error + Send + Sync + 'static) -> SerializationError {
+        SerializationError(Arc::new(err))
+    }
+
+    /// Retrieve an error reason by downcasting to specific type.
+    pub fn downcast_ref<T: Error + 'static>(&self) -> Option<&T> {
+        self.0.downcast_ref()
+    }
+
+    /// Try to get a mutable reference to the inner error by downcasting.
+    ///
+    /// This will only succeed if the `Arc` holding the error has no other
+    /// outstanding clones (i.e. the strong count is 1 and there are no weak
+    /// references).
+    pub fn try_downcast_mut<T: Error + 'static>(&mut self) -> Option<&mut T> {
+        Arc::get_mut(&mut self.0)?.downcast_mut()
+    }
+}
