+++ scylla/src/lib.rs
@@ -160,12 +160,12 @@ pub mod frame {
 // Note: When editing comment on submodules here edit corresponding comments
 // on scylla-cql modules too.
 pub mod serialize {
-    pub use scylla_cql::serialize::SerializationError;
+    pub use scylla_cql_core::serialize::SerializationError;
     /// Contains the [BatchValues][batch::BatchValues] and [BatchValuesIterator][batch::BatchValuesIterator] trait and their
     /// implementations.
     pub mod batch {
         // Main types
-        pub use scylla_cql::serialize::batch::{
+        pub use scylla_cql_core::serialize::batch::{
             BatchValues, BatchValuesFromIterator, BatchValuesIterator,
             BatchValuesIteratorFromIterator, TupleValuesIter,
         };
@@ -174,10 +174,10 @@ pub mod serialize {
     /// Contains the [SerializeRow][row::SerializeRow] trait and its implementations.
     pub mod row {
         // Main types
-        pub use scylla_cql::serialize::row::{RowSerializationContext, SerializeRow};
+        pub use scylla_cql_core::serialize::row::{RowSerializationContext, SerializeRow};
 
         // Errors
-        pub use scylla_cql::serialize::row::{
+        pub use scylla_cql_core::serialize::row::{
             BuiltinSerializationError, BuiltinSerializationErrorKind, BuiltinTypeCheckError,
             BuiltinTypeCheckErrorKind,
         };
@@ -186,10 +186,10 @@ pub mod serialize {
     /// Contains the [SerializeValue][value::SerializeValue] trait and its implementations.
     pub mod value {
         // Main types
-        pub use scylla_cql::serialize::value::SerializeValue;
+        pub use scylla_cql_core::serialize::value::SerializeValue;
 
         // Errors
-        pub use scylla_cql::serialize::value::{
+        pub use scylla_cql_core::serialize::value::{
             BuiltinSerializationError, BuiltinSerializationErrorKind, BuiltinTypeCheckError,
             BuiltinTypeCheckErrorKind, MapSerializationErrorKind, MapTypeCheckErrorKind,
             SetOrListSerializationErrorKind, SetOrListTypeCheckErrorKind,
@@ -200,7 +200,7 @@ pub mod serialize {
 
     /// Contains types and traits used for safe serialization of values for a CQL statement.
     pub mod writers {
-        pub use scylla_cql::serialize::writers::{
+        pub use scylla_cql_core::serialize::writers::{
             CellOverflowError, CellValueBuilder, CellWriter, RowWriter, WrittenCellProof,
         };
     }
